backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors==0.1
djangorestframework==3.16.1
sqlparse==0.5.3
